* Merged with otserv85

+++ otserv/trunk/condition.h
@@ -117,7 +117,7 @@ public:
 	virtual bool executeCondition(Creature* creature, int32_t interval);
 	virtual void endCondition(Creature* creature, ConditionEnd_t reason) = 0;
 	virtual void addCondition(Creature* creature, const Condition* condition) = 0;
-	virtual uint32_t getIcons() const;
+	virtual uint16_t getIcons() const;
 	ConditionId_t getId() const {return id;}
 	uint32_t getSubId() const {return subId;}
 
@@ -161,7 +161,7 @@ public:
 	virtual bool executeCondition(Creature* creature, int32_t interval);
 	virtual void endCondition(Creature* creature, ConditionEnd_t reason);
 	virtual void addCondition(Creature* creature, const Condition* condition);
-	virtual uint32_t getIcons() const;
+	virtual uint16_t getIcons() const;
 
 	virtual ConditionGeneric* clone()  const { return new ConditionGeneric(*this); }
 };
@@ -280,7 +280,7 @@ public:
 	virtual bool executeCondition(Creature* creature, int32_t interval);
 	virtual void endCondition(Creature* creature, ConditionEnd_t reason);
 	virtual void addCondition(Creature* creature, const Condition* condition);
-	virtual uint32_t getIcons() const;
+	virtual uint16_t getIcons() const;
 
 	virtual ConditionDamage* clone()  const { return new ConditionDamage(*this); }
 
@@ -326,7 +326,7 @@ public:
 	virtual bool executeCondition(Creature* creature, int32_t interval);
 	virtual void endCondition(Creature* creature, ConditionEnd_t reason);
 	virtual void addCondition(Creature* creature, const Condition* condition);
-	virtual uint32_t getIcons() const;
+	virtual uint16_t getIcons() const;
 
 	virtual ConditionSpeed* clone()  const { return new ConditionSpeed(*this); }
